backend/requirements.txt
Django>=3.1.4,<3.2.0
psycopg2-binary>=2.8.5,<2.9.0
django-choices>=1.7.0,<1.8.0
djangorestframework>=3.12.2,<3.13.0
